requirements.txt
fastapi
uvicorn[standard]
pydantic
chromadb>=0.5.0
numpy<2.0
minio
google-generativeai
pika
python-multipart
requests
sqlalchemy
psycopg2-binary
python-dotenv
alembic
mcp
redis
nest-asyncio
